move plugins/webapi -> packages/webapi
