+++ plugins/webapi/plugin.go
@@ -14,7 +14,8 @@ import (
 	"github.com/iotaledger/wasp/client"
 	"github.com/iotaledger/wasp/packages/parameters"
 	"github.com/iotaledger/wasp/packages/util/auth"
-	"github.com/iotaledger/wasp/plugins/webapi/httperrors"
+	"github.com/iotaledger/wasp/packages/webapi"
+	"github.com/iotaledger/wasp/packages/webapi/httperrors"
 	"github.com/labstack/echo/v4"
 	"github.com/labstack/echo/v4/middleware"
 	"github.com/pangpanglabs/echoswagger/v2"
@@ -59,7 +60,7 @@ func configure(*node.Plugin) {
 
 	auth.AddAuthentication(Server.Echo(), parameters.GetStringToString(parameters.WebAPIAuth))
 
-	addEndpoints(Server, adminWhitelist())
+	webapi.Init(Server, adminWhitelist())
 }
 
 func customHTTPErrorHandler(err error, c echo.Context) {
